partcipants activity created
