+++ app/src/main/res/layout/contact_list_item.xml
@@ -77,4 +77,19 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
+    <Button
+        android:id="@+id/btAdd"
+        android:visibility="gone"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="42dp"
+        android:layout_marginEnd="21dp"
+        android:minHeight="30dp"
+        android:background="@drawable/custom_button"
+        android:text="@string/add"
+        android:textAllCaps="false"
+        android:textColor="@drawable/text_color_button"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
 </androidx.constraintlayout.widget.ConstraintLayout>
